[#105] Feat: 패밀리 데이터 연동 수정 및 Suspense와 비슷하게 커스텀 구현
