+++ components/family/diary/family-diary-item.module.scss
