+++ components/family/growth/family-growth-list.module.scss
@@ -1,0 +1,9 @@
+.growth-list {
+  width: 100%;
+
+  & &__item {
+    width: 74px;
+    border-radius: 10px;
+    background-color: var(--main-color);
+  }
+}
