+++ components/family/growth/family-growth.module.scss
@@ -11,13 +11,3 @@
   font-weight: 700;
   color: #a9a9a9;
 }
-
-.growth-list {
-  width: 100%;
-
-  & &__item {
-    width: 74px;
-    border-radius: 10px;
-    background-color: var(--main-color);
-  }
-}
