+++ components/family/schedule/family-schedule-list.module.scss
@@ -1,0 +1,5 @@
+.schedule-list {
+  display: flex;
+  flex-direction: column;
+  gap: 10px 0;
+}
