+++ components/family/schedule/family-schedule-list.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import FamilyScheduleItem from '@/components/family/schedule/family-schedule-item';
+import styles from './family-schedule-list.module.scss';
+
+export default function FamilyScheduleList() {
+  return (
+    <ul className={styles['schedule-list']}>
+      <FamilyScheduleItem />
+      <FamilyScheduleItem />
+      <FamilyScheduleItem />
+    </ul>
+  );
+}
