+++ components/family/schedule/family-schedule.module.scss
@@ -11,9 +11,3 @@
   font-weight: 700;
   color: #a9a9a9;
 }
-
-.schedule-list {
-  display: flex;
-  flex-direction: column;
-  gap: 10px 0;
-}
